Made `email` a required field when creating a user
A mapping between an email address and a principal has been added which
allows us to resolve a user by their email address.
Email verification is NOT implemented yet.

This commit also moves `email` out of the options field when creating users
and makes it a required parameter. Lots of the tests that were using
createUser directly have been replaced with `TestsUtil.generateTestUsers`.